moved symbol tab to ast
