+++ inc/struct.h
@@ -120,16 +120,6 @@ bool hasValueExact(struct value *key, struct value *base);
  */
 bool hasValue(int key, struct value *base);
 
-/* = Symtab = */
-
-/** The symbol table which stores variables. */
-extern struct symbol symtab[NHASH];
-/** Searches the symbol table for a name and returns its location.
- * If the symbol table cannot find an already created version of the symbol, it will create a new entry and return it.
- * @param[in]  s The name of the symbol to search for.
- * @return   A pointer to the storage location of the symbol within the symbol table.
- */
-struct symbol *lookup(char *s);
 
 /* === MEMORY MANAGEMENT === */
